[NF] the ticket importer now supports versions and referneces, mantis is a reference implementation for that

+++ importer/mantis/src/main/scala/de/jowisoftware/mining/importer/mantis/ChangeParser.scala
@@ -142,7 +142,7 @@ class ChangeParser extends Logging {
   private def processRelationship(change: String) = change match {
     case relationshipRegex(relationString, ticketId) =>
       ValueUtils.relationshipStringToRelationshipType(relationString) match {
-        case Some(ticketRel) => TicketRelationship(ticketId, ticketRel)
+        case Some(ticketRel) => TicketRelationship(ticketId.toInt, ticketRel)
         case None => sys.error("Unparsable Relationship: "+relationString)
       }
     case _ => sys.error("Unparsable String: "+change)
